+++ src/main/java/com/vias/uc/backend/repository/AuditoriaRepository.java
@@ -3,4 +3,4 @@ package com.vias.uc.backend.repository;
 import com.vias.uc.backend.model.Auditoria;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuditoriaRepository extends JpaRepository<Auditoria, Long> {}
+public interface AuditoriaRepository extends JpaRepository<Auditoria, Integer> {}
